requirements.txt
customtkinter
tkinterdnd2
Pillow
httpx[http2]
requests-toolbelt
beautifulsoup4
keyring
loguru
pyperclip
pyinstaller
tenacity
pyyaml